removing extra database_functions file
